Refactor Grid class into hierarchy
